Rename to pyls_jsonrpc
