+++ MANIFEST.in
@@ -1,3 +1,3 @@
 include README.rst
 include versioneer.py
-include jsonrpc/_version.py
+include pyls_jsonrpc/_version.py
