+++ pyls_jsonrpc/_version.py
@@ -43,7 +43,7 @@ def get_config():
     cfg.style = "pep440"
     cfg.tag_prefix = ""
     cfg.parentdir_prefix = ""
-    cfg.versionfile_source = "jsonrpc/_version.py"
+    cfg.versionfile_source = "pyls_jsonrpc/_version.py"
     cfg.verbose = False
     return cfg
