+++ setup.cfg
@@ -1,7 +1,7 @@
 [versioneer]
 VCS = git
 style = pep440
-versionfile_source = jsonrpc/_version.py
-versionfile_build = jsonrpc/_version.py
+versionfile_source = pyls_jsonrpc/_version.py
+versionfile_build = pyls_jsonrpc/_version.py
 tag_prefix =
 parentdir_prefix =
